requirements/c9.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.3
django-cors-headers==2.1.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.10
openapi-codec==1.3.2
pytz==2017.2
requests==2.18.3
simplejson==3.11.1
uritemplate==3.0.0
urllib3==1.22
